requirements.txt
pydantic==1.9.2
uvicorn==0.18.3
fastapi==0.81.0
requests==2.29.0
ofxparse==0.21
Pillow==10.0.0